+++ pub/jaspr_unrouter/lib/jaspr_unrouter.dart
@@ -3,6 +3,7 @@ library;
 
 export 'src/core/route_data.dart';
 export 'src/core/route_definition.dart';
+export 'src/runtime/link.dart' show UnrouterLink, UnrouterLinkMode;
 export 'src/runtime/unrouter.dart';
 export 'src/runtime/navigation.dart'
     show UnrouterBuildContextExtension, UnrouterController, UnrouterScope;